# docs/auth/api/openapi.yaml
openapi: 3.0.3
info:
  title: Mainframe AI Assistant SSO API
  description: |
    Comprehensive Single Sign-On API for the Mainframe AI Assistant platform.

    This API provides secure authentication and authorization services supporting
    multiple identity providers including Google OAuth, Microsoft Entra ID, and SAML 2.0.

    ## Features
    - Multi-provider SSO support
    - JWT-based authentication
    - PKCE security for OAuth flows
    - Comprehensive audit logging
    - Rate limiting and security controls

    ## Security
    All endpoints implement industry-standard security practices including:
    - HTTPS-only communication
    - CSRF protection
    - Rate limiting
    - Input validation
    - Comprehensive logging

  version: 2.0.0
  contact:
    name: API Support
    email: api-support@yourcompany.com
    url: https://docs.yourcompany.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.yourcompany.com/v2
    description: Production server
  - url: https://staging-api.yourcompany.com/v2
    description: Staging server
  - url: http://localhost:3000/api/v2
    description: Development server
paths:
...
  /auth/sessions:
    get:
      summary: Get user sessions
      description: List active sessions for the current user
      operationId: getUserSessions
      tags:
        - Session Management
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: string
          description: Session identifier
        device_info:
          type: object
          properties:
            user_agent:
              type: string
              description: Browser/device user agent
            ip_address:
              type: string
              description: IP address
            location:
              type: string
              description: Approximate location
        created_at:
          type: string
          format: date-time
          description: Session creation time
        last_accessed:
          type: string
          format: date-time
          description: Last activity timestamp
        expires_at:
          type: string
          format: date-time
          description: Session expiration time
...